Internally, rename the metadata route to datasets, for consistency with it's actual on-site name
