+++ client/src/common_text/nav_lang.yaml
@@ -14,7 +14,7 @@ privacy_title:
   en: Privacy
   fr: Confidentialité
 
-metadata:
+datasets:
   en: Datasets
   fr: Données
