+++ client/src/panels/panel_declarations/covid/covid_expenditures.js
@@ -19,8 +19,8 @@ import { yearsWithCovidDataStore } from "src/models/covid/yearsWithCovidDataStor
 import { Dept } from "src/models/subjects";
 
 import { WrappedNivoPie } from "src/charts/wrapped_nivo/index";
+import { get_source_links } from "src/Datasets/utils";
 import { infographic_href_template } from "src/infographic/infographic_href_template";
-import { get_source_links } from "src/metadata/utils";
 import { minLargeDevice } from "src/style_constants/index";
 
 import {
