+++ client/src/panels/panel_declarations/results/result_drilldown/result_drilldown.js
@@ -15,7 +15,7 @@ import { LeafSpinner } from "src/components/index";
 
 import { ensure_loaded } from "src/core/ensure_loaded";
 
-import { get_source_links } from "src/metadata/utils";
+import { get_source_links } from "src/Datasets/utils";
 
 import ResultsExplorer from "./results_scheme";
