+++ client/src/panels/panel_declarations/services/provided_services_list.js
@@ -16,9 +16,8 @@ import {
   useServicesByProgram,
 } from "src/models/services/services_queries";
 
+import { get_source_links } from "src/Datasets/utils";
 import { infographic_href_template } from "src/infographic/infographic_href_template";
-
-import { get_source_links } from "src/metadata/utils";
 
 import text from "./services.yaml";
