+++ client/src/panels/panel_declarations/services/services_digital_status.js
@@ -22,7 +22,7 @@ import { is_a11y_mode } from "src/core/injected_build_constants";
 
 import { StandardLegend } from "src/charts/legends/index";
 import { WrappedNivoHBar } from "src/charts/wrapped_nivo/index";
-import { get_source_links } from "src/metadata/utils";
+import { get_source_links } from "src/Datasets/utils";
 import {
   secondaryColor,
   highlightOrangeColor,
