+++ client/src/panels/panel_declarations/services/services_intro.js
@@ -15,8 +15,8 @@ import {
   useServiceSummaryOrg,
 } from "src/models/services/services_queries";
 
+import { get_source_links } from "src/Datasets/utils";
 import { FormFrontend } from "src/FormFrontend";
-import { get_source_links } from "src/metadata/utils";
 
 // import { formats } from "src/core/format";
