+++ client/src/panels/panel_declarations/services/services_standards.js
@@ -19,7 +19,7 @@ import { is_a11y_mode } from "src/core/injected_build_constants";
 
 import Gauge from "src/charts/gauge";
 
-import { get_source_links } from "src/metadata/utils";
+import { get_source_links } from "src/Datasets/utils";
 
 import text from "./services.yaml";
 import "./services.scss";
